requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coverage==7.5.1
Django==4.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.0.0
idna==3.7
inflection==0.5.1
mccabe==0.7.0
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.1
sqlparse==0.4.4
stripe==9.7.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
